Add reducer tests for account recovery state tree. Also, extract the dummy data to `test-data.js` so different test suites can share it.
