+++ client/state/account-recovery/actions.js
@@ -7,6 +7,7 @@ import wpcom from 'lib/wp';
  * Internal dependencies
  */
 import {
+	ACCOUNT_RECOVERY_FETCH,
 	ACCOUNT_RECOVERY_FETCH_SUCCESS,
 	ACCOUNT_RECOVERY_FETCH_FAILED,
 } from 'state/action-types';
@@ -25,15 +26,16 @@ export const accountRecoveryFetchFailed = ( error ) => {
 	};
 };
 
-export const accountRecoveryFetch = () => {
-	return ( dispatch ) => {
-		wpcom.undocumented().me().getAccountRecovery( ( error, accountRecoverySettings ) => {
-			if ( error ) {
-				dispatch( accountRecoveryFetchFailed( error ) );
-				return;
-			}
+export const accountRecoveryFetch = () => ( dispatch ) => {
+	dispatch( { type: ACCOUNT_RECOVERY_FETCH } );
 
-			dispatch( accountRecoveryFetchSuccess( accountRecoverySettings ) );
+	return new Promise( ( resolve, reject ) => {
+		wpcom.undocumented().me().getAccountRecovery( ( error, accountRecoverySettings ) => {
+			error ? reject( error ) : resolve( accountRecoverySettings );
 		} );
-	};
+	} ).then( ( accountRecoverySettings ) => {
+		dispatch( accountRecoveryFetchSuccess( accountRecoverySettings ) );
+	} ).catch( ( error ) => {
+		dispatch( accountRecoveryFetchFailed( error ) );
+	} );
 };
