+++ client/state/account-recovery/test/actions.js
@@ -21,19 +21,9 @@ import {
 	ACCOUNT_RECOVERY_FETCH_FAILED,
 } from '../../action-types';
 
-describe( 'account-recovery ctions', () => {
-	const dummyData = {
-		email: 'dummytest@a8ctest.com',
-		email_validated: false,
-		phone: {
-			country_code: 'TW',
-			country_numeric_code: '+886',
-			number: '0918000000',
-			number_full: '+886918000000',
-		},
-		phone_validated: false,
-	};
+import dummyData from './test-data';
 
+describe( 'account-recovery ctions', () => {
 	const spy = sinon.spy();
 
 	beforeEach( () => {
