+++ client/state/account-recovery/test/test-data.js
@@ -1,0 +1,13 @@
+const dummyData = {
+	email: 'dummytest@a8ctest.com',
+	email_validated: false,
+	phone: {
+		country_code: 'TW',
+		country_numeric_code: '+886',
+		number: '0918000000',
+		number_full: '+886918000000',
+	},
+	phone_validated: false,
+};
+
+export default dummyData;
